Добавлена кнопка оплаты, убрано лишнее поле при заказе
